requirements.txt
streamlit==1.28.0
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
huggingface_hub==0.19.1 